Made the solver for linear equations over F2 search for a "short" vector to return. Updated DA arrows for both arcslide and antibraid.
Note the setting of maximum size of coeffs_a in autocompleteda. The
settings used to produce the current results are:

arcslide: maximum length is 3 in all cases.

antibraid, short: maximum length is 2 in all cases.

antibraid, long: maximum length is 2 in the first two rounds, and 4 in
the last round (outside regions).